requirement.txt
streamlit
langchain-openai
langchain-huggingface
faiss-cpu
numpy
python-dotenv